Eliminate duplicate code in watch command tests
- Created shared helper runFileRemovalTest for file removal watch tests
- Refactored watch-file-removal.test.ts and watch-file-removal-change.test.ts to use helper
- Created shared helper setupWatchChangeTest for watch change event tests
- Refactored watch-duplicate-change.test.ts and watch-unknown-change.test.ts to use helper
- Reduced code duplication from 64 lines (675 tokens) to 0
