+++ src/cli/test/test-helpers/watch-file-removal.ts
@@ -1,0 +1,84 @@
+import type { WatchListener } from "node:fs";
+import { mkdtemp, rm, writeFile } from "node:fs/promises";
+import { tmpdir } from "node:os";
+import path from "node:path";
+
+import { runWatchCommand } from "../../src/commands/watch.js";
+import { DependencyTracker } from "../../src/modules/dependency-tracker.js";
+import { withTemporaryProperty } from "./temporary-property.js";
+import { createMockWatchFactory } from "./watch-fixtures.js";
+
+export interface FileRemovalTestOptions {
+    tmpPrefix: string;
+    eventType: "change" | "rename";
+}
+
+export interface FileRemovalTestResult {
+    removedFilePath: string | null;
+}
+
+/**
+ * Shared test helper for file removal scenarios in watch command.
+ * Sets up a temporary directory with a test file, configures dependency tracking,
+ * removes the file, and triggers the specified watch event.
+ */
+export async function runFileRemovalTest(options: FileRemovalTestOptions): Promise<FileRemovalTestResult> {
+    const root = await mkdtemp(path.join(tmpdir(), options.tmpPrefix));
+    const targetFile = path.join(root, "script1.gml");
+    await writeFile(targetFile, "var x = 1;", "utf8");
+
+    const abortController = new AbortController();
+    const listenerCapture: { listener: WatchListener<string> | undefined } = { listener: undefined };
+    const removeFileDescriptor = Object.getOwnPropertyDescriptor(DependencyTracker.prototype, "removeFile");
+    if (!removeFileDescriptor || typeof removeFileDescriptor.value !== "function") {
+        throw new Error("Expected DependencyTracker.removeFile to be defined");
+    }
+    const callOriginalRemoveFile = Function.prototype.call.bind(removeFileDescriptor.value) as (
+        tracker: DependencyTracker,
+        filePath: string
+    ) => void;
+    let removedFilePath: string | null = null;
+    let resolveRemoval: (() => void) | null = null;
+    const removalPromise = new Promise<void>((resolve) => {
+        resolveRemoval = resolve;
+    });
+
+    const watchFactory = createMockWatchFactory(listenerCapture);
+
+    await withTemporaryProperty(
+        DependencyTracker.prototype,
+        "removeFile",
+        function (this: DependencyTracker, filePath: string): void {
+            removedFilePath = filePath;
+            resolveRemoval?.();
+            return callOriginalRemoveFile(this, filePath);
+        },
+        async () => {
+            const watchPromise = runWatchCommand(root, {
+                extensions: [".gml"],
+                verbose: false,
+                quiet: true,
+                debounceDelay: 0,
+                runtimeServer: false,
+                websocketServer: false,
+                statusServer: false,
+                abortSignal: abortController.signal,
+                watchFactory
+            });
+
+            await new Promise((resolve) => setTimeout(resolve, 50));
+
+            await rm(targetFile, { force: true });
+            listenerCapture.listener?.(options.eventType, path.basename(targetFile));
+
+            await Promise.race([removalPromise, new Promise((resolve) => setTimeout(resolve, 500))]);
+
+            abortController.abort();
+            await watchPromise;
+        }
+    );
+
+    await rm(root, { recursive: true, force: true });
+
+    return { removedFilePath };
+}
